refactor: better code structure
