+++ packages/maskbook/src/extension/options-page/DashboardComponents/WalletAssetsTable.tsx
@@ -19,8 +19,8 @@ import classNames from 'classnames'
 import { useStylesExtends } from '../../../components/custom-ui-helper'
 import { formatBalance, formatCurrency } from '../../../plugins/Wallet/formatter'
 import { useI18N } from '../../../utils/i18n-next-ui'
-import { CurrencyType, AssetDetailed, ERC20TokenDetailed, EthereumTokenType } from '../../../web3/types'
-import { getTokenUSDValue, isSameAddress } from '../../../web3/helpers'
+import { CurrencyType, ERC20TokenDetailed, EthereumTokenType } from '../../../web3/types'
+import { isSameAddress } from '../../../web3/helpers'
 import { TokenIcon } from './TokenIcon'
 import type { WalletRecord } from '../../../plugins/Wallet/database/types'
 import { ActionsBarFT } from './ActionsBarFT'
@@ -29,6 +29,8 @@ import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
 import { useAssetsDetailed } from '../../../web3/hooks/useAssetsDetailed'
 import { useTrustedERC20TokensFromDB } from '../../../plugins/Wallet/hooks/useERC20Tokens'
 import { useStableTokensDebank } from '../../../web3/hooks/useStableTokensDebank'
+import type { AssetDetailed } from '../../../plugins/Wallet/types'
+import { getTokenUSDValue } from '../../../plugins/Wallet/helpers'
 
 const MAX_TOKENS_LENGTH = 5
 const MIN_VALUE = 5
