+++ packages/maskbook/src/plugins/Wallet/services/assets.ts
@@ -1,10 +1,16 @@
 import { values } from 'lodash-es'
 import BigNumber from 'bignumber.js'
-import { AssetProvider, ZerionAddressAsset } from '../types'
+import { EthereumAddress } from 'wallet.ts'
+import { AssetDetailed, AssetProvider, BalanceRecord, PortfolioProvider, ZerionAddressAsset } from '../types'
 import * as OpenSeaAPI from '../apis/opensea'
 import * as ZerionAPI from '../apis/zerion'
-import type { AssetDetailed } from '../../../web3/types'
+import * as DebankAPI from '../apis/debank'
+import { CurrencyType } from '../../../web3/types'
 import { ChainId, EthereumTokenType } from '../../../web3/types'
+import { getChainId } from '../../../extension/background-script/EthereumService'
+import { unreachable } from '../../../utils/utils'
+import { createEtherToken } from '../../../web3/helpers'
+import { formatChecksumAddress } from '../formatter'
 
 export async function getAssetsListNFT(address: string, provider: AssetProvider) {
     if (provider === AssetProvider.OPENSEAN) {
@@ -26,16 +32,55 @@ export async function getAssetsListNFT(address: string, provider: AssetProvider)
     return []
 }
 
-const filterAssetType = ['compound', 'trash', 'uniswap', 'uniswap-v2', 'nft']
-
-// TOOD:
-// unify asset from different assets provider
-export async function getAssetsList(address: string): Promise<AssetDetailed[]> {
-    const { meta, payload } = await ZerionAPI.getAssetsList(address)
-    if (meta.status !== 'ok') throw new Error('Fail to load assets.')
-    const assetsList = values(payload.assets)
-    return formatAssetsFromZerion(assetsList)
+export async function getAssetsList(address: string, provider: PortfolioProvider): Promise<AssetDetailed[]> {
+    if (!EthereumAddress.isValid(address)) return []
+    switch (provider) {
+        case PortfolioProvider.ZERION:
+            const { meta, payload } = await ZerionAPI.getAssetsList(address)
+            if (meta.status !== 'ok') throw new Error('Fail to load assets.')
+            const assetsList = values(payload.assets)
+            return formatAssetsFromZerion(assetsList)
+        case PortfolioProvider.DEBANK:
+            if ((await getChainId()) !== ChainId.Mainnet) return []
+            const { data = [], error_code } = await DebankAPI.getAssetsList(address)
+            if (error_code === 0) return formatAssetsFromDebank(data)
+            return []
+        default:
+            unreachable(provider)
+    }
 }
+
+function formatAssetsFromDebank(data: BalanceRecord[]) {
+    return data.map(
+        (x): AssetDetailed => ({
+            chain: x.chain,
+            token:
+                x.id === 'eth'
+                    ? createEtherToken(ChainId.Mainnet)
+                    : {
+                          // distinguish token type
+                          type: EthereumTokenType.ERC20,
+                          address: formatChecksumAddress(x.id),
+                          chainId: ChainId.Mainnet,
+                          name: x.name,
+                          symbol: x.symbol,
+                          decimals: x.decimals,
+                      },
+            balance: new BigNumber(x.balance).toFixed(),
+            price: {
+                [CurrencyType.USD]: new BigNumber(x.price).toFixed(),
+            },
+            value: {
+                [CurrencyType.USD]: new BigNumber(x.price)
+                    .multipliedBy(new BigNumber(x.balance).dividedBy(new BigNumber(10).pow(x.decimals)))
+                    .toFixed(),
+            },
+            logoURL: x.logo_url,
+        }),
+    )
+}
+
+const filterAssetType = ['compound', 'trash', 'uniswap', 'uniswap-v2', 'nft']
 
 function formatAssetsFromZerion(data: ZerionAddressAsset[]) {
     return data
