+++ packages/maskbook/src/web3/hooks/useStableTokensDebank.ts
@@ -1,14 +1,14 @@
 import { useAsyncRetry } from 'react-use'
 import { formatChecksumAddress } from '../../plugins/Wallet/formatter'
+import type { BalanceListResponse } from '../../plugins/Wallet/types'
 import { createERC20Token } from '../helpers'
 import { ChainId } from '../types'
-import type { Debank } from './useAssetsDetailedDebank'
 import { useChainId } from './useChainState'
 
 async function fetcher(chainId: ChainId) {
     if (chainId !== ChainId.Mainnet) return []
     const response = await fetch('https://api.debank.com/stable_coins')
-    const { data = [], error_code } = (await response.json()) as Debank.BalanceListResponse
+    const { data = [], error_code } = (await response.json()) as BalanceListResponse
     if (error_code === 0) return data
     return []
 }
